Edited delete street method

+++ App/controllers/admin.py
@@ -48,8 +48,11 @@ def admin_delete_area(area_id):
     db.session.delete(area)
     db.session.commit()
 
-def admin_delete_street(street_id):
-    street = Street.query.get(street_id)
+def admin_delete_street(area_id, street_id):
+    area = Area.query.get(area_id)
+    if not area:
+        raise ValueError("Invalid area ID.")
+    street = Street.query.filter_by(areaId=area_id).get(street_id)
     if not street:
         raise ValueError("Invalid street ID.")
     db.session.delete(street)
